+++ src/hints.js
@@ -15,14 +15,14 @@ class Hints {
       range,
       renderOptions: {
         before: {
-          opacity: 0.4,
+          opacity: config.get('opacity'),
           color: new ThemeColor('phpParameterHint.hintForeground'),
           contentText: message,
           backgroundColor: new ThemeColor('phpParameterHint.hintBackground'),
           margin: `0px ${config.get('margin') + 1}px 0px ${config.get(
             'margin'
           )}px;padding: ${config.get('verticalPadding')}px ${config.get('horizontalPadding')}px;`,
-          borderRadius: '5px',
+          borderRadius: `${config.get('borderRadius')}px`,
           fontStyle: config.get('fontStyle'),
           fontWeight: `${config.get('fontWeight')};font-size:${config.get('fontSize')}px;`
         }